Set tab to inactive color when the contact is in inactive state
Fixes the disparity between various Psi modes. In non-tabbed mode,
the "(Inactive)" suffix is added to the window title. In tabbed mode,
there was no suffix. However, adding the suffix alone was not enough,
because in the all-in-one mode there is no window title displayed.

Also, Gone status is now regarded as inactive too.