Rely on QNetworkSession to decide when to connect after resume.
This quickens connection by connecting right after dhcp finished but probably may have some side effects like
choosing wrong network configuration for connection establishment. For example when we have ethernet but
vpn is not yet connected. Or we have ethernet connected but only wifi has access to the Internet.
In any case Qt has a concept of connection roaming which should help in this case, but it's not yet tested.