Fix cppcheck notes with severity "warning":
Member variable 'ActiveProfiles::d' is not initialized in the constructor.
Member variable 'AvTransmitHandler::previousThread' is not initialized in the constructor.
Member variable 'MCmdSimpleSite::cursorPos' is not initialized in the constructor.
Member variable 'ChatViewThemeProvider::_psi' is not initialized in the constructor.
Member variable 'ChatViewPrivate::*' is not initialized in the constructor.
Member variable 'ToolbarPrefs::dirty' is not initialized in the constructor.
Member variable 'GroupMenu::createNewGroupAction_' is not initialized in the constructor.
Member variable 'Thread::result_' is not initialized in the constructor.
Member variable 'HoverableStyleOptionViewItem::hovered' is not initialized in the constructor.
Member variable 'JingleVoiceCaller::*' is not initialized in the constructor.
Member variable 'MCmdSimpleState::flags_' is not initialized in the constructor.

+ Small code refactoring.