Fix cppcheck notes with severity "warning":
Member variable 'SeparatorAction::d' is not initialized in the constructor.
Member variable 'MUCItemsTask::action_' is not initialized in the constructor.
Member variable 'NetworkAccessManager::_handlerSeed' is not initialized in the constructor.
Member variable 'OptionsTabAppearanceMisc::parentWidget' is not initialized in the constructor.
Member variable 'OptionsTabAppearanceGeneral::*' is not initialized in the constructor.
Member variable 'OptionsTabApplication::*' is not initialized in the constructor.
Member variable 'OptionsTabGroupchat::dlg' is not initialized in the constructor.
Member variable 'OptionsTabPlugins::psi' is not initialized in the constructor.
Member variable 'OptionsTabSound::parentWidget' is not initialized in the constructor.
Member variable 'OptionsTabStatusAuto::parentWidget' is not initialized in the constructor.
Member variable 'OptionsTabStatusGeneral::*' is not initialized in the constructor.
Member variable 'OptionsTabToolbars::*' is not initialized in the constructor.
Member variable 'PrivacyListItem::*' is not initialized in the constructor.