Fix cppcheck notes with severity "warning":
Member variable 'PsiContactList::deinitAccCounter_' is not initialized in the constructor.
Member variable 'AuthEvent::v_nick' is not initialized in the constructor.
Member variable 'ServerListQuerier::redirectCount_' is not initialized in the constructor.
Member variable 'SxeEdit::null_' is not initialized in the constructor.
Member variable 'SxeRecord::*' is not initialized in the constructor.
Member variable 'TabCompletion::*' is not initialized in the constructor.
Member variable 'UserResource::v_pgpVerifyStatus' is not initialized in the constructor.
Member variable 'IconsetSelectItem::*' is not initialized in the constructor.
Member variable 'IconsetDisplayItem::*' is not initialized in the constructor.
Member variable 'PsiTabBar::*' is not initialized in the constructor.
Member variable 'PsiTipLabel::*' is not initialized in the constructor.

+ Code refactoring.