Fixed crash in range-based loop for QHash
No idea what was wrong QHash iterator but it crashed on ++ operation. This first happened on replacing foreach with range-based for. According to docs it had to work. There is some positive experience for example here https://www.kdab.com/qt-range-based-for-loops-and-structured-bindings/ and according to C++ standards docs our invocation was safe (no temporaries generating our temporaries)