Change runtime linking path instead of compile-time one
The earlier commit on this branch incorrectly assumed GCC would get
the hint from the -L option that /usr/X11R7/lib might need
inclusion at runtime linking. This is not the case, rather it needs
to see a -Wl option to pass the parameters directly to the runtime
linker.

Signed-off-by: devfonks <nkeck72@netzero.net>