Make the build reproducible
Whilst working on the Reproducible Builds effort [0] I noticed that
psi could not be built reproducibly.

This is due to using CMake's TIMESTAMP feature without specifying a timezone.
This means that, even if SOURCE_DATE_EPOCH is present, the value (and thus the
binary) will vary depending on the build machine's timezone setting.

I originally filed this in Debian as bug #1017473 [1].

 [0] https://reproducible-builds.org/
 [1] https://bugs.debian.org/1017473